Added function in ArticleService to delete article if it exists. Now we are able to run tests multiple times.

+++ Article/src/main/java/de/leuphana/article/structure/database/ArticleDatabase.java
@@ -3,7 +3,7 @@ package de.leuphana.article.structure.database;
 import de.leuphana.article.structure.database.entity.ArticleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleDatabase extends JpaRepository<ArticleEntity, Long> {
+public interface ArticleDatabase extends JpaRepository<ArticleEntity, Integer> {
 
     ArticleEntity findArticleEntityByName(String name);
